feat(data/list/defs): Add fold*_with_index
We add two folding functions for lists, `foldl_with_index` and
`foldr_with_index`. They generalise `foldl` and `foldr`, respectively,
by passing the index of the current element to the combining function.
This gives us more concise implementations of `find_indexes` and
`indexes_values`. We also add two new functions which are implemented in
terms of `foldr_with_index`: `map_with_index` (like `map`, but with
indexes) and `index_list` (the list of indexes of a list).